requirements.txt
ffmpeg-python
librosa>=0.10.0
soundfile
pydub
pyloudnorm
numpy
opencv-python
Pillow
tqdm
edge-tts
google-generativeai
